Finsh script added
Added a finishscript so you get another angle if you go over the finish
