+++ ProjectSettings/TagManager.asset
@@ -4,7 +4,6 @@
 TagManager:
   serializedVersion: 2
   tags:
-  - Boost
   - Ground
   layers:
   - Default
